Fix singularity in R6
R6 has a 1/(delta -1) term in Delta_dd() which of course is singular
when delta = 1 (i.e. when the density is the critical density). It is
considered an artefact, resulting from the mathematical form of phi_r(),
which ha no physical sense. So a fix is to shift sslightly delta when it
is equal to 1. This is what this commit does.